refactor(src): lint code, separate library root
