+++ .flake8
@@ -1,0 +1,10 @@
+[flake8]
+max-line-length = 99
+select = C,E,F,W,B,B950
+ignore = E501,W503,E203
+exclude =
+    .git
+    .mypy_cache
+    static
+    dist
+    *.egg-info
