+++ .importlinter
@@ -1,0 +1,9 @@
+[importlinter]
+root_package = _garnet
+include_external_packages = true
+
+[importlinter:contract:1]
+name = Private package can not import from public package
+type = forbidden
+source_modules = _garnet
+forbidden_modules = garnet
