+++ .isort.cfg
@@ -1,0 +1,7 @@
+[settings]
+multi_line_output = 3
+include_trailing_comma = true
+force_grid_wrap = 0
+use_parentheses = true
+line_length = 99
+known_third_party = ["telethon"]
