--- garnet/_client/__init__.py
@@ -1,1 +1,0 @@
-from ._client import TelegramClient
