--- garnet/events/__init__.py
@@ -1,4 +1,0 @@
-from .dispatcher import EventDispatcher
-from .filter import Filter
-from .handler import EventHandler
-from .router import Router
