--- garnet/filters/__init__.py
@@ -1,2 +1,0 @@
-from . import text
-from .file_ext import File
