--- garnet/storages/__init__.py
@@ -1,4 +1,0 @@
-# Source: https://github.com/aiogram/aiogram
-
-from .json import JSONStorage
-from .memory import MemoryStorage
