--- garnet/vars/fsm.py
@@ -1,5 +1,0 @@
-import contextvars
-
-from garnet.storages.base import FSMContext
-
-State: contextvars.ContextVar[FSMContext] = contextvars.ContextVar("State")
