+++ pyproject.toml
@@ -14,7 +14,8 @@ classifiers = [
     "Programming Language :: Python :: 3.8",
 ]
 packages = [
-    {include = "garnet"},
+    {include = "_garnet", from = "src"},
+    {include = "garnet", from = "src"}
 ]
 
 [tool.poetry.dependencies]
@@ -26,9 +27,27 @@ aioredis = {version = "^1.3.1", optional = true}
 aioredis = "^1.3.1"
 black = "^19.10b0"
 ipython = "^7.15.0"
+import-linter = "^1.1"
+flake8 = "^3.8.3"
+isort = {version = "^5.5.2", extras = ["pyproject"]}
 
 [tool.poetry.extras]
 redis = ["aioredis"]
+
+[tool.black]
+line-length = 99
+target-version = ['py37', 'py38']
+exclude = '''
+(
+    \.eggs
+  | \.git
+  | \.tox
+  | build
+  | dist
+  | venv
+  | docs
+)
+'''
 
 [build-system]
 requires = ["poetry>=0.12"]
