+++ src/_garnet/filters/state.py
@@ -1,0 +1,47 @@
+from typing import Container, Union
+
+from _garnet.events.filter import Filter
+from _garnet.vars import fsm
+
+ANY_STATE = (id(all), id(any))
+
+
+class _MetaCurrentState(type):
+    @classmethod
+    def __eq__(mcs, _s: Union[str, Container[str]]) -> Filter:
+        if id(_s) in ANY_STATE:
+
+            def _f(_):
+                return True
+
+        elif not isinstance(_s, str) and isinstance(_s, Container):
+
+            async def _f(_):
+                return (await fsm.StateCtx.get().get_state()) in _s
+
+        else:
+
+            async def _f(_):
+                return (await fsm.StateCtx.get().get_state()) == _s
+
+        return Filter(_f, None)
+
+    @classmethod
+    def exact(mcs, state: Union[str, Container[str]]) -> Filter:
+        # noinspection PyTypeChecker
+        return State == state  # type: ignore
+
+
+class State(metaclass=_MetaCurrentState):
+    """
+    Class with implemented in metaclass magic methods.
+    Any state: ANY_STATE states "*". Simply use equality operator to builtin methods
+        CurrentState == all
+        CurrentState == any
+    Particular state(s):
+        CurrentState == ["state_1", "state_2"]
+        CurrentState == "state_1"
+    """
+
+
+__all__ = ("State",)
