+++ src/_garnet/vars/fsm.py
@@ -1,0 +1,7 @@
+import contextvars
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from _garnet.events.fsm_context import FSMContext
+
+StateCtx: "contextvars.ContextVar[FSMContext]" = contextvars.ContextVar("StateCtx")
