+++ src/garnet/storages.py
@@ -1,0 +1,9 @@
+from _garnet.storages.base import BaseStorage
+from _garnet.storages.json import JSONStorage
+from _garnet.storages.memory import MemoryStorage
+
+__all__ = (
+    "BaseStorage",
+    "JSONStorage",
+    "MemoryStorage",
+)
